@@ -1,21 +1,35 @@
 "use client";
 import { useEffect, useState } from "react";
+import QuestionForm, { QuestionType } from "./QuestionForm";
+import CsvImportGuide from "./CsvImportGuide";
 
 type Module = { id: string; order: number; title: string };
-type Question = { id: string; text: string; options: string[]; correctIndex: number; active: boolean };
+type Question = { 
+  id: string; 
+  text: string; 
+  options: string[]; 
+  correctIndex: number; 
+  active: boolean;
+  questionType?: string;
+  optionA?: string;
+  optionB?: string;
+  optionC?: string;
+  optionD?: string;
+  correctAnswer?: string;
+};
 
 export default function AdminQuestions() {
   const [modules, setModules] = useState<Module[]>([]);
   const [moduleId, setModuleId] = useState<string>("");
   const [questions, setQuestions] = useState<Question[]>([]);
-  const [text, setText] = useState("");
-  const [opts, setOpts] = useState<string[]>(["", "", "", ""]);
-  const [correctIndex, setCorrectIndex] = useState(0);
   const [message, setMessage] = useState<string | null>(null);
   const [editingId, setEditingId] = useState<string | null>(null);
   const [eText, setEText] = useState("");
   const [eOptions, setEOptions] = useState("");
   const [eCorrectIndex, setECorrectIndex] = useState(0);
+  const [eQuestionType, setEQuestionType] = useState<QuestionType>("MCQ_4");
+  const [uploading, setUploading] = useState(false);
+  const [importMsg, setImportMsg] = useState<string | null>(null);
 
   async function loadModules() {
     const res = await fetch("/api/admin/modules", { cache: "no-store" });
@@ -33,12 +47,34 @@ export default function AdminQuestions() {
   }
   useEffect(() => { loadQuestions(); }, [moduleId]);
 
-  async function addQuestion(e: React.FormEvent) {
-    e.preventDefault(); setMessage(null);
-    const res = await fetch(`/api/admin/questions`, { method: "POST", headers: { "Content-Type": "application/json" }, body: JSON.stringify({ moduleId, text, options: opts, correctIndex }) });
-    const data = await res.json().catch(() => ({}));
-    if (!res.ok) { setMessage(data.message || "Failed to add"); return; }
-    setText(""); setOpts(["", "", "", ""]); setCorrectIndex(0);
+  async function addQuestion(data: {
+    text: string;
+    options: string[];
+    correctIndex: number;
+    questionType: QuestionType;
+    optionA?: string;
+    optionB?: string;
+    optionC?: string;
+    optionD?: string;
+    correctAnswer: string;
+  }) {
+    setMessage(null);
+    const res = await fetch(`/api/admin/questions`, { 
+      method: "POST", 
+      headers: { "Content-Type": "application/json" }, 
+      body: JSON.stringify({ 
+        moduleId, 
+        ...data
+      }) 
+    });
+    
+    const responseData = await res.json().catch(() => ({}));
+    
+    if (!res.ok) { 
+      setMessage(responseData.message || "Failed to add"); 
+      return; 
+    }
+    
     loadQuestions();
   }
 
@@ -62,12 +98,33 @@ export default function AdminQuestions() {
     setEText(q.text);
     setEOptions(q.options.join(" | "));
     setECorrectIndex(q.correctIndex);
+    setEQuestionType((q.questionType as QuestionType) || "MCQ_4");
   }
-  async function saveEdit() {
+
+  async function saveEdit(data: {
+    text: string;
+    options: string[];
+    correctIndex: number;
+    questionType: QuestionType;
+    optionA?: string;
+    optionB?: string;
+    optionC?: string;
+    optionD?: string;
+    correctAnswer: string;
+  }) {
     if (!editingId) return;
-    const options = eOptions.split("|").map((s) => s.trim()).filter(Boolean);
-    const res = await fetch(`/api/admin/questions/${editingId}`, { method: "PATCH", headers: { "Content-Type": "application/json" }, body: JSON.stringify({ text: eText, options, correctIndex: eCorrectIndex }) });
-    if (!res.ok) { setMessage("Save failed"); return; }
+    
+    const res = await fetch(`/api/admin/questions/${editingId}`, { 
+      method: "PATCH", 
+      headers: { "Content-Type": "application/json" }, 
+      body: JSON.stringify(data) 
+    });
+    
+    if (!res.ok) { 
+      setMessage("Save failed"); 
+      return; 
+    }
+    
     setEditingId(null);
     loadQuestions();
   }
@@ -75,6 +132,28 @@ export default function AdminQuestions() {
   return (
     <main>
       <h1 className="text-2xl font-semibold mb-4 text-[color:var(--color-brand)]">Questions</h1>
+      <section className="rounded border border-slate-200 bg-white p-4 shadow-[var(--shadow-card)] my-4">
+        <h3 className="font-semibold mb-2">Bulk Import (CSV)</h3>
+        <CsvImportGuide />
+        <form onSubmit={async (e) => {
+          e.preventDefault(); setImportMsg(null); setUploading(true);
+          const input = document.getElementById('csvfile') as HTMLInputElement | null;
+          if (!input || !input.files || input.files.length === 0) { setImportMsg('Choose a CSV file'); setUploading(false); return; }
+          const fd = new FormData(); fd.append('file', input.files[0]);
+          const res = await fetch('/api/admin/questions/import', { method: 'POST', body: fd });
+          const data = await res.json().catch(() => ({}));
+          if (!res.ok) { setImportMsg(data.message || 'Import failed'); setUploading(false); return; }
+          setImportMsg(`Imported ${data.imported}/${data.total}. Failed: ${data.failed}`);
+          setUploading(false);
+          loadQuestions();
+        }} className="grid gap-2 max-w-[640px]">
+          <input id="csvfile" type="file" accept=".csv,text/csv" />
+          <div className="flex items-center gap-2">
+            <button disabled={uploading} className="rounded bg-[color:var(--color-brand)] text-white px-3 py-2 text-sm hover:opacity-95 disabled:opacity-50">{uploading ? 'Uploading…' : 'Upload & Import'}</button>
+            {importMsg && <span className="text-sm text-slate-700">{importMsg}</span>}
+          </div>
+        </form>
+      </section>
       <div className="mb-2">
         <label className="mr-2">Module:</label>
         <select className="rounded border border-slate-300 px-3 py-2 text-sm" value={moduleId} onChange={(e) => setModuleId(e.target.value)}>
@@ -86,16 +165,7 @@ export default function AdminQuestions() {
 
       <section className="rounded border border-slate-200 bg-white p-4 shadow-[var(--shadow-card)] my-4">
         <h3 className="font-semibold mb-2">Add Question</h3>
-        <form onSubmit={addQuestion} className="grid gap-3 max-w-[640px]">
-          <textarea className="w-full rounded border border-slate-300 px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-[color:var(--color-brand)]" placeholder="Question text" value={text} onChange={(e) => setText(e.target.value)} required />
-          {opts.map((o, i) => (
-            <div key={i} className="flex items-center gap-2">
-              <input type="radio" name="correct" checked={correctIndex === i} onChange={() => setCorrectIndex(i)} />
-              <input className="w-full rounded border border-slate-300 px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-[color:var(--color-brand)]" placeholder={`Option ${i + 1}`} value={o} onChange={(e) => setOpts((arr) => arr.map((v, idx) => (idx === i ? e.target.value : v)))} required />
-            </div>
-          ))}
-          <button type="submit" className="rounded bg-[color:var(--color-brand)] text-white px-3 py-2 text-sm hover:opacity-95">Add</button>
-        </form>
+        <QuestionForm onSubmit={addQuestion} />
       </section>
 
       <h3 className="font-semibold mb-2">Existing</h3>
@@ -104,6 +174,7 @@ export default function AdminQuestions() {
           <thead>
             <tr>
               <th className="text-left border-b border-slate-200 px-2 py-2 text-sm font-semibold">Text</th>
+              <th className="text-left border-b border-slate-200 px-2 py-2 text-sm font-semibold">Type</th>
               <th className="text-left border-b border-slate-200 px-2 py-2 text-sm font-semibold">Options</th>
               <th className="text-left border-b border-slate-200 px-2 py-2 text-sm font-semibold">Correct</th>
               <th className="text-left border-b border-slate-200 px-2 py-2 text-sm font-semibold">Active</th>
@@ -114,31 +185,21 @@ export default function AdminQuestions() {
             {questions.map((q) => (
               <tr key={q.id}>
                 <td className="px-2 py-2 border-b border-slate-100">
-                  {editingId === q.id ? (
-                    <textarea className="w-full rounded border border-slate-300 px-2 py-1 text-sm" value={eText} onChange={(e) => setEText(e.target.value)} />
-                  ) : (
-                    q.text
-                  )}
+                  {q.text}
                 </td>
                 <td className="px-2 py-2 border-b border-slate-100">
-                  {editingId === q.id ? (
-                    <input className="w-full rounded border border-slate-300 px-2 py-1 text-sm" value={eOptions} onChange={(e) => setEOptions(e.target.value)} placeholder="Separate with |" />
-                  ) : (
-                    q.options.join(" | ")
-                  )}
+                  {q.questionType || "MCQ_4"}
                 </td>
                 <td className="px-2 py-2 border-b border-slate-100">
-                  {editingId === q.id ? (
-                    <input className="w-24 rounded border border-slate-300 px-2 py-1 text-sm" type="number" min={0} value={eCorrectIndex} onChange={(e) => setECorrectIndex(Number(e.target.value))} />
-                  ) : (
-                    q.options[q.correctIndex]
-                  )}
+                  {q.options.join(" | ")}
+                </td>
+                <td className="px-2 py-2 border-b border-slate-100">
+                  {q.correctAnswer || q.options[q.correctIndex]}
                 </td>
                 <td className="px-2 py-2 border-b border-slate-100">{q.active ? "Yes" : "No"}</td>
                 <td className="px-2 py-2 border-b border-slate-100">
                   {editingId === q.id ? (
                     <div className="flex gap-2">
-                      <button className="rounded bg-[color:var(--color-brand)] text-white px-3 py-1.5 text-sm hover:opacity-95" onClick={saveEdit}>Save</button>
                       <button className="rounded border border-slate-300 bg-white px-3 py-1.5 text-sm hover:bg-slate-50" onClick={() => setEditingId(null)}>Cancel</button>
                     </div>
                   ) : (
@@ -154,6 +215,29 @@ export default function AdminQuestions() {
           </tbody>
         </table>
       </div>
+      
+      {/* Edit Question Modal */}
+      {editingId && (
+        <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-50">
+          <div className="bg-white rounded-lg p-6 max-w-2xl w-full max-h-[90vh] overflow-y-auto">
+            <h3 className="text-xl font-semibold mb-4">Edit Question</h3>
+            <QuestionForm 
+              onSubmit={saveEdit}
+              initialText={eText}
+              initialOptions={eOptions.split("|").map(s => s.trim()).filter(Boolean)}
+              initialCorrectIndex={eCorrectIndex}
+              initialQuestionType={(questions.find(q => q.id === editingId)?.questionType as QuestionType) || "MCQ_4"}
+              buttonText="Save Changes"
+            />
+            <button 
+              className="mt-4 w-full rounded border border-slate-300 bg-white px-3 py-2 text-sm hover:bg-slate-50"
+              onClick={() => setEditingId(null)}
+            >
+              Cancel
+            </button>
+          </div>
+        </div>
+      )}
     </main>
   );
 }
