@@ -50,7 +50,7 @@ export async function POST(req: Request) {
   if (!user) return NextResponse.json({ message: "User not found" }, { status: 404 });
   const score = await computeOverallScore(userId);
   const filePath = await generateCertificatePdf({ userName: user.name, userEmail: user.email, overallScore: score });
-  await prisma.certificate.upsert({ where: { userId }, update: { filePath, totalScore: score }, create: { userId, filePath, totalScore: score } });
+  await (prisma as any).certificate.upsert({ where: { userId_mainModuleId: { userId, mainModuleId: null } }, update: { filePath, totalScore: score }, create: { userId, mainModuleId: null, filePath, totalScore: score } });
   const url = `/api/certificate/download?userId=${encodeURIComponent(userId)}`;
   return NextResponse.json({ url });
 }
