@@ -6,12 +6,11 @@ import { useSearchParams } from "next/navigation";
 export default function Login() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
-  const [otp, setOtp] = useState("");
   const sp = useSearchParams();
   const err = sp.get("error");
   let errMsg: string | null = null;
   if (err === "CredentialsSignin") {
-    errMsg = "Sign-in failed. Check email/password. If admin MFA is enabled, enter your 6-digit OTP. Also ensure your email is verified or reset your password if locked.";
+    errMsg = "Sign-in failed. Check email/password. Please ensure your email is verified or reset your password if locked.";
   } else if (err) {
     errMsg = "Sign-in error. Please try again.";
   }
@@ -32,10 +31,12 @@ export default function Login() {
         <div className="grid gap-3 text-center">
           <input className="w-full rounded border border-slate-300 px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-[color:var(--color-brand)]" placeholder="email" value={email} onChange={e=>setEmail(e.target.value)} />
           <input className="w-full rounded border border-slate-300 px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-[color:var(--color-brand)]" placeholder="password" type="password" value={password} onChange={e=>setPassword(e.target.value)} />
-          <input className="w-full rounded border border-slate-300 px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-[color:var(--color-brand)]" placeholder="OTP (if required)" value={otp} onChange={e=>setOtp(e.target.value)} />
           <button className="rounded bg-[color:var(--color-brand)] text-white px-3 py-2 text-sm hover:opacity-95" onClick={async()=>{
-            const body: any = { email, password, callbackUrl:"/" };
-            if (otp.trim()) body.otp = otp.trim();
+            const body = { 
+              email, 
+              password, 
+              callbackUrl:"/" 
+            };
             await signIn("credentials", body);
           }}>Sign In</button>
         </div>
